Changed the timer thread implementation under Win32.
Timer events are still generated with timerSetEvent.
However, the RtLab timer-loop code runs in another
separate thread synchronized to the timer by means of
a condition variable. In this way if the RtLab code
does not finish within the loop period it will catch-up
to the next availiable timer event. Lost events are counted.
TODO: report the lost events in RtTimerLoop::stat()